comment mapper entity에 통합 #84
