+++ src/main/java/com/tsp/new_tsp_admin/api/domain/comment/AdminCommentDTO.java
@@ -16,8 +16,8 @@ import javax.validation.constraints.NotEmpty;
 @AllArgsConstructor
 @ApiModel(value = "코멘트 관련 변수")
 public class AdminCommentDTO extends NewCommonDTO {
-    @ApiModelProperty(required = true, value = "rnum", hidden = true, example = "1")
-    private Integer rnum;
+    @ApiModelProperty(required = true, value = "rowNum", hidden = true, example = "1")
+    private Integer rowNum;
 
     @ApiModelProperty(required = true, value = "idx", hidden = true, example = "1")
     private Long idx;
