+++ src/test/java/com/tsp/new_tsp_admin/api/comment/service/AdminCommentJpaServiceTest.java
@@ -1,6 +1,5 @@
 package com.tsp.new_tsp_admin.api.comment.service;
 
-import com.tsp.new_tsp_admin.api.comment.mapper.AdminCommentMapper;
 import com.tsp.new_tsp_admin.api.domain.comment.AdminCommentDTO;
 import com.tsp.new_tsp_admin.api.domain.comment.AdminCommentEntity;
 import com.tsp.new_tsp_admin.api.domain.model.AdminModelEntity;
@@ -81,7 +80,7 @@ class AdminCommentJpaServiceTest {
                 .visible("Y")
                 .build();
 
-        adminCommentDTO = AdminCommentMapper.INSTANCE.toDto(adminCommentEntity);
+        adminCommentDTO = AdminCommentEntity.toDto(adminCommentEntity);
     }
 
     @BeforeEach
@@ -185,7 +184,7 @@ class AdminCommentJpaServiceTest {
                 .visible("Y")
                 .build();
 
-        adminCommentDTO = AdminCommentMapper.INSTANCE.toDto(adminCommentEntity);
+        adminCommentDTO = AdminCommentEntity.toDto(adminCommentEntity);
 
         // when
         when(mockAdminCommentJpaService.findOneAdminComment(adminCommentEntity)).thenReturn(adminCommentDTO);
@@ -218,7 +217,7 @@ class AdminCommentJpaServiceTest {
                 .visible("Y")
                 .build();
 
-        adminCommentDTO = AdminCommentMapper.INSTANCE.toDto(adminCommentEntity);
+        adminCommentDTO = AdminCommentEntity.toDto(adminCommentEntity);
 
         // when
         given(mockAdminCommentJpaService.findOneAdminComment(adminCommentEntity)).willReturn(adminCommentDTO);
@@ -295,7 +294,7 @@ class AdminCommentJpaServiceTest {
                 .visible("Y")
                 .build();
 
-        AdminCommentDTO adminCommentDTO = AdminCommentMapper.INSTANCE.toDto(adminCommentEntity);
+        AdminCommentDTO adminCommentDTO = AdminCommentEntity.toDto(adminCommentEntity);
 
         adminCommentJpaService.updateAdminComment(adminCommentEntity);
 
@@ -331,7 +330,7 @@ class AdminCommentJpaServiceTest {
                 .visible("Y")
                 .build();
 
-        AdminCommentDTO adminCommentDTO = AdminCommentMapper.INSTANCE.toDto(adminCommentEntity);
+        AdminCommentDTO adminCommentDTO = AdminCommentEntity.toDto(adminCommentEntity);
 
         adminCommentJpaService.updateAdminComment(adminCommentEntity);
 
@@ -368,7 +367,7 @@ class AdminCommentJpaServiceTest {
     void 어드민코멘트삭제Mockito테스트() throws Exception {
         // given
         adminCommentJpaService.insertAdminComment(adminCommentEntity);
-        adminCommentDTO = AdminCommentMapper.INSTANCE.toDto(adminCommentEntity);
+        adminCommentDTO = AdminCommentEntity.toDto(adminCommentEntity);
 
         // when
         when(mockAdminCommentJpaService.findOneAdminComment(adminCommentEntity)).thenReturn(adminCommentDTO);
@@ -391,7 +390,7 @@ class AdminCommentJpaServiceTest {
     void 어드민코멘트삭제BDD테스트() throws Exception {
         // given
         adminCommentJpaService.insertAdminComment(adminCommentEntity);
-        adminCommentDTO = AdminCommentMapper.INSTANCE.toDto(adminCommentEntity);
+        adminCommentDTO = AdminCommentEntity.toDto(adminCommentEntity);
 
         // when
         given(mockAdminCommentJpaService.findOneAdminComment(adminCommentEntity)).willReturn(adminCommentDTO);
